requirements.txt
streamlit
streamlit-drawable-canvas
tensorflow==2.13.0
numpy
pillow
opencv-python
